pyproject.toml
[project]
name = "google-flights-agent"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
requires-python = ">=3.14,<4.0"
dependencies = [
    "beautifulsoup4 (>=4.14.3,<5.0.0)",
    "python-telegram-bot (>=22.7,<23.0)",
    "loguru (>=0.7.3,<0.8.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
